oops... II - The Second Coming

+++ src/Jaket/IO/Writer.cs
@@ -88,7 +88,7 @@ public class Writer
 
         if (LobbyController.IsCurrentMultikillLobby())
         {
-            Marshal.WriteInt32(mem, Inc(4), (short)((weapon << 25) | (Convert.ToByte(team) << 17) | (emoji << 13) | (rps << 10) | ((typing ? 1 : 0) << 9)));
+            Marshal.WriteInt32(mem, Inc(4), (weapon << 25) | (Convert.ToByte(team) << 17) | (emoji << 13) | (rps << 10) | ((typing ? 1 : 0) << 9));
             return;
         }
